Compiling and working with u16 MAPSEC
Plenty of changes; struct was being padded/aligned automatically, but the generated .inc files do not account for this change in struct size

+++ src/daycare.c
@@ -1651,7 +1651,7 @@ static void AddHatchedMonToParty(u8 id)
     u8 name[12];
     u16 ball;
     u16 caughtLvl;
-    u8 mapNameID;
+    u16 mapNameID;
     struct Pokemon* mon = &gPlayerParty[id];
 
     CreatedHatchedMon(mon, &gEnemyParty[0]);
